==> flight_predictions/notebooks/4.0-dyk-Weather 2017.ipynb <==
import pandas as pd
import numpy as np
import glob
import os
import re, operator, warnings
warnings.filterwarnings('ignore')

#import the weather files for top 20 origin airports of 2018 obtained from:
# https://mesonet.agron.iastate.edu/ASOS/
path = '../data/raw/2017/'
flight_path = '../data/interim/'
all_files = glob.glob(os.path.join(path, "*.csv"))     # advisable to use os.path.join as this makes concatenation OS independent

df_from_each_file = (pd.read_csv(f, parse_dates=['valid']) for f in all_files)
concatenated_df = pd.concat(df_from_each_file)

weather_df = concatenated_df

weather_df.columns

weather_df = weather_df[['station','valid', 'tmpf','vsby','gust','sknt','p01i',
                         'ice_accretion_3hr','ice_accretion_6hr',]]

weather_df.columns

weather_df = weather_df.fillna(0)

weather_df['p01i'] = pd.to_numeric(weather_df.p01i, errors='coerce')

weather_df['valid_new'] = weather_df.valid.apply(lambda x: x.replace(minute=0))

df1 = pd.DataFrame(weather_df.groupby(['station','valid_new'])['tmpf','vsby'].min())
df2 = pd.DataFrame(weather_df.groupby(['station','valid_new'])['p01i'].max())
df3 = pd.DataFrame(weather_df.groupby(['station','valid_new'])['gust','sknt'].max())

df_test = pd.merge(df1, df2, how='left', left_index=True, right_index=True)

df_test = df_test.merge(df3, how='left', left_index=True, right_index=True)

df_test.head()

df_test.reset_index(inplace=True) 

df_test.head()

weather_df_new = df_test.copy().fillna(0)

weather_df_new.head()

#export raw, concoctanated weather file for 2017 top 20 airports
weather_df_new.to_csv(f'{path}/2017WeatherALL.csv', index = False, encoding = 'utf-8')

#import the 2017 Flight Data and clean it
df = pd.read_csv(f'{flight_path}/2017ALL.csv', parse_dates=['FlightDate'])


df = df.dropna(subset=['Tail_Number','DepTime','DepDelay','DepartureDelayGroups','TaxiOut',
                      'WheelsOff','WheelsOn','TaxiIn','ArrTime','ArrDelay','CRSElapsedTime',
                      'ActualElapsedTime','AirTime'])

df = df.fillna(0)

#Let's remove the features that will not be available for predicting flights
df = df[['Quarter', 'Month', 'DayofMonth', 'DayOfWeek', 'FlightDate',
       'Reporting_Airline', 'Tail_Number', 'Flight_Number_Reporting_Airline',
       'Origin', 'OriginState', 'DestState', 'Dest', 'CRSDepTime', 'CRSArrTime', 'CRSElapsedTime',
       'ArrDelay', 'Distance', 'DistanceGroup',]]
df = df.reindex()

#since the DEP_TIME is presented as a float we need to convert it to a Timestamp
def convert_to_time (datatype):
    '''This method will take a float as input and returns a datetime.time object'''
    to_string = str(int(datatype))
    #check length of string and add leading zeros if lenght is less than 4
    while len(to_string) < 4:
        to_string = '0' + to_string
    #check if hour starts with 24 then change to 00
    if to_string[:2] == '24':
        to_string = '00' + to_string[2:]
    #obj = time.strptime(to_string, '%H%M')
    
    to_string = to_string[:2] + ':' + to_string[2:] 
    datetime = pd.to_datetime(to_string)
    datetime = pd.to_datetime(to_string, format='%H:%M')
    time = datetime.time()
    return time
    
    #datetime = time.strftime("%H:%M", obj)
    #return datetime
    
#We now need to create a new Timestamp column incorporating the information from FL_DATE and DEP_TIME
def combine_datetime (dataframe):
    '''This function will extract needed data from two columns and combine into a single Timestamp object'''
    year = dataframe[0].year
    month = dataframe[0].month
    day = dataframe[0].day
    hour = dataframe[1].hour
    minute = dataframe[1].minute
    second = dataframe[1].second
    
    datetime = pd.Timestamp(year, month, day, hour, minute, second)
    return datetime

combined = df[['FlightDate','CRSDepTime']]

#convert DEP_TIME to timestamp
combined['CRSDepTime'] = df.CRSDepTime.apply(convert_to_time)

#apply the combine_datetime function to optimized_df and create a new column 'FL_TIME'
df['FL_TIME'] = combined.apply(combine_datetime, axis=1)

df['FL_TIME_NEW'] = df.FL_TIME.apply(lambda x: x.replace(minute=0))

df.head()

weather_df_new1 = weather_df_new.copy()

#remove rows will null values which is approximately 1.5%
null_columns = weather_df_new1.columns[weather_df_new1.isnull().any()]
weather_df_new1[null_columns].isnull().sum()

weather_df_new.head()

#combine the station name and datetime for easier / faster merging with flight data
weather_df_new1 = pd.DataFrame()
weather_df_new1['combined'] = weather_df_new['station'] + weather_df_new['valid_new'].astype(str)
weather_df_new1['precip'] = weather_df_new['p01i']
weather_df_new1['vis'] = weather_df_new['vsby']
weather_df_new1['temp'] = weather_df_new['tmpf']
weather_df_new1['wind'] = weather_df_new['sknt']

weather_df_new1.head()

#suming up the precipitation for last five hours
sum_precip = []
for row in range(len(weather_df_new1.precip) - 5):
    rows_sum = sum(weather_df_new1.iloc[row:row + 5, 1].values)
    sum_precip.append(rows_sum)

sum_precip.insert(0,0)
sum_precip.insert(0,0)
sum_precip.insert(0,0)
sum_precip.insert(0,0)
sum_precip.insert(0,0)

#Averaging the visability for last five hours
ave_vis = []
for row in range(len(weather_df_new1.vis) - 5):
    rows_ave = np.mean(weather_df_new1.iloc[row:row + 5, 1].values)
    ave_vis.append(rows_ave)

ave_vis.insert(0,0)
ave_vis.insert(0,0)
ave_vis.insert(0,0)
ave_vis.insert(0,0)
ave_vis.insert(0,0)

weather_df_new1['ave_vis'] = ave_vis

weather_df_new1['precip_sum'] = sum_precip

weather_df_new1 = weather_df_new1.drop(columns=['precip','vis'])

weather_df_new1.head()

#do the same in the flight data 
df['combined'] =  df['Origin'] + df['FL_TIME_NEW'].astype(str)

df.head()

copy_df = df.copy()

copy_df.head()

weather_df_new1.head()

#mapping the weather data to the flight dataframe

weather_df_new1['zip'] = list(zip(weather_df_new1.temp, weather_df_new1.wind, weather_df_new1.ave_vis,
                                  weather_df_new1.precip_sum))

weather_df_new1.head()

look_up = weather_df_new1.copy().drop(columns=['temp','wind','ave_vis','precip_sum'])

look_up = look_up.set_index('combined')

look_up.head()

weather_dict = look_up.to_dict()['zip']

copy_df['zipped'] = copy_df.combined.map(weather_dict)

copy_df = copy_df.fillna(0)

copy_df['zipped'] = copy_df.zipped.apply(lambda x: (0,0,0,0) if x==0 else x)

 copy_df[['tempF','wind','ave_vis','precip_sum']] = pd.DataFrame(copy_df['zipped'].tolist(), index=df.index)                                                                                                                       

copy_df.head()

#export the dataframe as a csv to model
copy_df.to_csv(f'{flight_path}2017NEW.csv', index = False, encoding = 'utf-8')
