==> flight_predictions/notebooks/6.0-dyk-modeling.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import os, re, operator, warnings
warnings.filterwarnings('ignore')
import pickle

#Define the location of the Data
path = '../data/'

# Import Feature Engineered Dataset

df = pd.read_csv(f'{path}/interim/2017FeatEng.csv')

df.head()

df.shape

#create a list with top 20 airport origins by flight volume
top_20 = list(df.groupby('Origin').size().sort_values(ascending=False)[0:20].index)
#create a list with top 5 airlines by flight volume
top_5 = list(df.groupby('Reporting_Airline').size().sort_values(ascending=False)[0:5].index)
#only use top 20 origin airports
top_df = df[df.Origin.isin(top_20)]
#only use top 5 airlines
top_df = top_df[top_df.Reporting_Airline.isin(top_5)]

top_df.shape

#Add a column classifying if the flight is delayed or not using 15 minutes as delayed
top_df['STATUS'] = top_df.ArrDelay.apply(lambda x: 1 if x >= 15 else 0)

# Label Encoding Categorical Values

from sklearn.preprocessing import LabelEncoder

#Ensure the columns are strings
top_df['mf_name'] = top_df['mf_name'].astype(str)
top_df['mf_year'] = top_df['mf_year'].astype(str)
top_df['plane_model'] = top_df['plane_model'].astype(str)
top_df['eng_model'] = top_df['eng_model'].astype(str)

le = LabelEncoder()
top_df['Reporting_Airline'] = le.fit_transform(top_df.Reporting_Airline.values)
top_df['Tail_Number'] = le.fit_transform(top_df.Tail_Number.values)
top_df['Origin'] = le.fit_transform(top_df.Origin.values)
top_df['Dest'] = le.fit_transform(top_df.Dest.values)
top_df['DOM_DIRECTION'] = le.fit_transform(top_df.DOM_DIRECTION.values)
top_df['OriginState'] = le.fit_transform(top_df.OriginState.values)
top_df['DestState'] = le.fit_transform(top_df.DestState.values)
top_df['mf_name'] = le.fit_transform(top_df.mf_name.values)
top_df['mf_year'] = le.fit_transform(top_df.mf_year.values)
top_df['plane_model'] = le.fit_transform(top_df.plane_model.values)
top_df['eng_model'] = le.fit_transform(top_df.eng_model.values)

top_df = top_df.drop(columns='FlightDate')

top_df.head()

# Looking at the data Imbalance

down = top_df[top_df.STATUS == 1]
up = top_df[top_df.STATUS == 0]
down = down.Reporting_Airline.count()
up = up.Reporting_Airline.count()
print(f'Delay percentage: {down/(up+down)}')

# Resampling the data to deal with Imbalance

from sklearn.utils import resample

#let's first separate majority class and minority class and resample

df_majority = top_df[top_df.STATUS == 0]
df_minority = top_df[top_df.STATUS == 1]


# Downsample majority class
df_majority_downsampled = resample(df_majority, 
                                 replace=False,    # sample without replacement
                                 n_samples=down,     # to match minority class
                                 random_state=42) # reproducible results
# combine the new dataframes
df_downsampled = pd.concat([df_majority_downsampled, df_minority])

df_downsampled.STATUS.value_counts()

# Splitting the Data into Train / Test Split

df_downsampled.columns

df_downsampled.describe()

df_downsampled['tempF'] = df_downsampled.tempF.apply(lambda x: (x - 32)*(5/9) + 273.15)

df_downsampled['plane_age'] = df_downsampled.plane_age.apply(lambda x: 0 if x < 0 else x)

from sklearn.model_selection import train_test_split
y = df_downsampled.STATUS.values
X = df_downsampled.drop(columns=['STATUS', 'ArrDelay']).values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state=42)

#clean up memory
del df, top_df

# Modeling (untuned)

## Neural Network Classifier

from sklearn.neural_network import MLPClassifier

mlp = MLPClassifier(random_state=seed)

mlp.fit(X_train, y_train)
y_pred_mlp = mlp.predict(X_test)
# Evaluate
accuracy = accuracy_score(y_test, y_pred_mlp)
precision = precision_score(y_test, y_pred_mlp)
recall = recall_score(y_test, y_pred_mlp)
cm = confusion_matrix(y_test, y_pred_mlp)
# Evaluate clf's accuracy on the test set
print(f'MLP: accuracy: {accuracy}, precision: {precision},\
          recall: {recall}, confusion matrix: {cm}')

parameter_space = {
    'hidden_layer_sizes': [(50,50,50), (50,100,50), (100,)],
    'activation': ['tanh', 'relu'],
    'solver': ['sgd', 'adam'],
    'alpha': [0.0001, 0.05],
    'learning_rate': ['constant','adaptive'],
}

from sklearn.model_selection import GridSearchCV

clf_mlp = GridSearchCV(mlp, parameter_space, n_jobs=-1, cv=3)
clf_mlp.fit(X_train, y_train)

# Best paramete set
print('Best parameters found:\n', clf_mlp.best_params_)

# All results
means = clf_mlp.cv_results_['mean_test_score']
stds = clf_mlp.cv_results_['std_test_score']
for mean, std, params in zip(means, stds, clf_mlp.cv_results_['params']):
    print("%0.3f (+/-%0.03f) for %r" % (mean, std * 2, params))

y_true, y_pred = y_test, clf_mlp.predict(X_test)

from sklearn.metrics import classification_report
print('Results on the test set:')
print(classification_report(y_true, y_pred))

## LightGBM

import lightgbm as lgb
params = {'learning_rate':0.01,
          'objective' :'binary',
          'num_leaves' : 100,
          'feature_fraction': 0.75, 
          'bagging_fraction': 0.8, 
          'bagging_freq':1,
          'boosting_type' : 'gbdt',
          'metric': 'binary_logloss'}
train_data = lgb.Dataset(X_train, label=y_train)

lgbm = lgb.train(params, train_data, 100)

y_pred_lgb = lgbm.predict(X_test)

y_pred_lgb

y_binary = np.where(y_pred_lgb < 0.5, 0, 1)

print(classification_report(y_test, y_binary))

# Evaluate
accuracy = accuracy_score(y_test, y_binary)
precision = precision_score(y_test, y_binary)
recall = recall_score(y_test, y_binary)
cm = confusion_matrix(y_test, y_binary)
# Evaluate clf's accuracy on the test set
print(f'LightGBM: accuracy: {accuracy}, precision: {precision},\
        recall: {recall}, confusion matrix: {cm}')

## Scikit-Learn Classifiers and XGBoost

#Import classifiers
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, ExtraTreesClassifier
from xgboost import XGBClassifier

#instantiate models
seed = 42
clf = RandomForestClassifier(random_state=seed)
xgb = XGBClassifier(random_state=seed)
ada = AdaBoostClassifier(random_state=seed)
xtr = ExtraTreesClassifier(random_state=seed)

classifiers = [('Random Forest', clf), ('XGBoost', xgb), ('ExtraTrees', xtr)]

from sklearn.metrics import confusion_matrix, accuracy_score,\
precision_score, recall_score, precision_recall_curve, classification_report

# Iterate over the pre-defined list of classifiers
for clf_name, clf_algo in classifiers:    
 
    # Fit clf to the training set
    clf_algo.fit(X_train, y_train)    
   
    # Predict y_pred
    y_pred = clf_algo.predict(X_test)
    
    # Evaluate
    accuracy = accuracy_score(y_test, y_pred)
    precision = precision_score(y_test, y_pred)
    recall = recall_score(y_test, y_pred)
    cm = confusion_matrix(y_test, y_pred)
   
    # Evaluate clf's accuracy on the test set
    print(f'name: {clf_name}')
    print(classification_report(y_test, y_pred))
    # Evaluate clf's accuracy on the test set 
    print(f'accuracy: {accuracy}')
    print(f'precision: {precision}')
    print(f'recall: {recall}')
    print(f'confusion matrix: {cm}')
    print('------------------------------')

# Import VotingClassifier from sklearn.ensemble
from sklearn.ensemble import VotingClassifier

# Instantiate a VotingClassifier vc
vc = VotingClassifier(estimators=classifiers)     

# Fit vc to the training set
vc.fit(X_train, y_train)   

# Evaluate the test set predictions
y_pred = vc.predict(X_test)

# Calculate accuracy score

accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred)
recall = recall_score(y_test, y_pred)
cm = confusion_matrix(y_test, y_pred)


# Evaluate clf's accuracy on the test set
print(f'name: Voting Classifier')
print(classification_report(y_test, y_pred))
# Evaluate clf's accuracy on the test set 
print(f'accuracy: {accuracy}')
print(f'precision: {precision}')
print(f'recall: {recall}')
print(f'confusion matrix: {cm}')
print('------------------------------')

# Tuning

from sklearn.model_selection import RandomizedSearchCV
from pprint import pprint

# Number of trees for tree ensambles
n_estimators = [int(x) for x in np.linspace(start = 100, stop = 150, num = 2)]
# Number of features to consider at every split
max_features = ['auto', 'sqrt']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(50, 105, num = 5)]
max_depth.append(None)
# Minimum number of samples required to split a node
min_samples_split = [5, 10, 20]
# Minimum number of samples required at each leaf node
min_samples_leaf = [6, 8, 12]
# Method of selecting samples for training each tree
bootstrap = [True, False]
# Create the random grid

## Random Forest

# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}
pprint(random_grid)

# Use the random grid to search for best hyperparameters
# First create the base model to tune
# Random search of parameters, using 3 fold cross validation, 
rf_random = RandomizedSearchCV(estimator = clf, param_distributions = random_grid, n_iter = 20, \
                               cv = 3, verbose=2, random_state=42, n_jobs = -1)
# Fit the random search model
rf_random.fit(X_train, y_train)

rf_random.best_params_

### Save the best params

#define and store ExtraTreeClassifier best parameters
rf_params = rf_random.best_params_
with open('rf_params', 'wb') as f:
    pickle.dump(rf_params, f)

## XGBoost

parameters = {'nthread':[4], #when use hyperthread, xgboost may become slower
              #'objective':['reg:linear'],
              'learning_rate': [0.01, 0.1],
              'max_depth': [6, 12], 
              'min_child_weight': [2, 6],
              #'verbosity': [1],
              'subsample': [1.0],
              'colsample_bytree': [0.3, 0.5],
              'gamma': [0],
              'n_estimators': [100, 150]}

xgb_rand = RandomizedSearchCV(xgb,
                              parameters,
                              cv = 3,
                              n_jobs = 4,
                              verbose=True)

xgb_rand.fit(X_train, y_train)

xgb_rand.best_params_

### Save the best params

#define and store ExtraTreeClassifier best parameters
xgb_params = xgb_rand.best_params_
with open('xgb_params', 'wb') as f:
    pickle.dump(xgb_params, f)

## ExtraTrees

pprint(xtr.get_params)

param_dist = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap
             }

xtr_tune = RandomizedSearchCV(estimator = xtr, param_distributions = param_dist, n_iter = 50, \
                               cv = 3, verbose=2, random_state=42, n_jobs = -1)
# Fit the random search model
xtr_tune.fit(X_train, y_train)
xtr_tune.best_params_

### Save the best params

#define and store ExtraTreeClassifier best parameters
xtr_params = xtr_tune.best_params_
with open('xtr_params', 'wb') as f:
    pickle.dump(xtr_params, f)

# Loading best parameters

Random Forest

#import Random Forest best parameters
with open('rf_params', 'rb') as f:
    rf_params = pickle.load(f)

XGBoost

#import XGBoost best parameters
with open('xgb_params', 'rb') as f:
    xgb_params = pickle.load(f)

ExtraTreeClassifier

#import ExtraTreeClassifier best parameters
with open('xtr_params', 'rb') as f:
    xtr_params = pickle.load(f)

# Modeling (tuned)

#instantiate models
seed = 42
new_clf = RandomForestClassifier(**rf_random.best_params_, random_state=seed)
new_xgb = XGBClassifier(**xgb_rand.best_params_, random_state=seed)
new_xtr = ExtraTreesClassifier(**xtr_tune.best_params_, random_state=seed)

classifiers = [('Random Forest', new_clf), ('XGBoost', new_xgb), ('ExtraTrees', new_xtr)]

# Iterate over the pre-defined list of classifiers
for clf_name, clf_algo in classifiers:    

    # Fit clf to the training set
    clf_algo.fit(X_train, y_train)    
    
    # Predict y_pred
    y_pred = clf_algo.predict(X_test)
          
    # Evaluate
    accuracy = accuracy_score(y_test, y_pred)
    precision = precision_score(y_test, y_pred)
    recall = recall_score(y_test, y_pred)
    cm = confusion_matrix(y_test, y_pred)
    print(f'name: {clf_name}')
    print(classification_report(y_test, y_pred))
    # Evaluate clf's accuracy on the test set 
    print(f'accuracy: {accuracy}')
    print(f'precision: {precision}')
    print(f'recall: {recall}')
    print(f'confusion matrix: {cm}')

# Instantiate a VotingClassifier vc
vc = VotingClassifier(estimators=classifiers)     

# Fit vc to the training set
vc.fit(X_train, y_train)   

# Evaluate the test set predictions
y_pred = vc.predict(X_test)

# Calculate accuracy score

accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred)
recall = recall_score(y_test, y_pred)
cm = confusion_matrix(y_test, y_pred)
print(f'Voting Classifier: ')
print(classification_report(y_test, y_pred))
    # Evaluate clf's accuracy on the test set 
print(f'accuracy: {accuracy}')
print(f'precision: {precision}')
print(f'recall: {recall}')
print(f'confusion matrix: {cm}')

# Results Summary

## Untuned

 - Random Forest
 - accuracy: 0.6444
 - precision: 0.6708
 - recall: 0.5697
 - confusion matrix: [[65281 25466]
 [39193 51901]]

 
 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |65281|25466|
 |Actual Delayed |39193|51901|

 - XGBoost: 
 - accuracy: 0.6436
 - precision: 0.6502
 - recall: 0.6245
 - confusion matrix: [[60151 30596]
 [34205 56889]]

 
 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |60151|30596|
 |Actual Delayed |34205|56889|

 - Extremely Randomized Trees: 
 - accuracy: 0.6424
 - precision: 0.6659
 - recall: 0.5744
 - confusion matrix: [[64499 26248]
 [38767 52327]]

 
 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |64499|26248|
 |Actual Delayed |38767|52327|

## Tuned

 - Random Forest, 
 - accuracy: 0.6856
 - precision: 0.6950
 - recall: 0.6635
 - confusion matrix: [[64226 26521]
 [30647 60447]]

 
 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |64226|26521|
 |Actual Delayed |30647|60447|

 - XGBoost, 
 - accuracy: 0.6894
 - precision: 0.6983
 - recall: 0.6689
 - confusion matrix: [[64428 26319]
 [30156 60938]]

 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |64428|26319|
 |Actual Delayed |30156|60938|

 - Extremely Randomized Trees
 - accuracy: 0.6824
 - precision: 0.6895
 - recall: 0.6657
 - confusion matrix: [[63448 27299]
 [30449 60645]]

 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |63448|27299|
 |Actual Delayed |30449|60645|

## Voting Classifier

 - Voting Classifier:
 - accuracy: 0.6892
 - precision: 0.6978
 - recall: 0.6694
 - confusion matrix: [[64345 26402]
 [30112 60982]]

 |' | Predicted On-time | Predicted Delayed |
 |-- | --- | --- |
 |Actual On-time |64345|26402|
 |Actual Delayed |30112|60982|
